requirements.txt
django>=4.2
pillow
psycopg2-binary
djangorestframework
djangorestframework-gis
